revert weakdeps
