+++ src/AutoDiff.jl
@@ -1,3 +1,12 @@
+module AutoDiff
+
+using ..ParametricMCPs: ParametricMCPs, get_problem_size, get_result_buffer, get_parameter_dimension
+using ChainRulesCore: ChainRulesCore
+using EnzymeCore: EnzymeCore, EnzymeRules
+using ForwardDiff: ForwardDiff
+using SparseArrays: SparseArrays
+using LinearAlgebra: LinearAlgebra
+
 function _solve_jacobian_θ(problem, solution, θ; active_tolerance = 1e-3)
     (; jacobian_z!, jacobian_θ!, lower_bounds, upper_bounds) = problem
     z_star = solution.z
@@ -38,9 +47,9 @@ function _solve_jacobian_θ(problem, solution, θ; active_tolerance = 1e-3)
 end
 
 function EnzymeRules.forward(
-    func::EnzymeCore.Const{typeof(solve)},
+    func::EnzymeCore.Const{typeof(ParametricMCPs.solve)},
     ::Type{ReturnType},
-    problem::EnzymeCore.Annotation{<:ParametricMCP},
+    problem::EnzymeCore.Annotation{<:ParametricMCPs.ParametricMCP},
     θ::EnzymeCore.Annotation;
     kwargs...,
 ) where {ReturnType}
@@ -52,11 +61,11 @@ function EnzymeRules.forward(
     end
 
     # backward pass
-    ∂z∂θ = ParametricMCPs._solve_jacobian_θ(problem.val, solution_val, θ.val)
+    ∂z∂θ_thunk = _solve_jacobian_θ(problem.val, solution_val, θ.val)
 
     if ReturnType <: EnzymeCore.BatchDuplicated
         solution_dval = map(θ.dval) do θdval
-            (; solution_val..., z = ∂z∂θ * θdval)
+            (; solution_val..., z = ∂z∂θ_thunk * θdval)
         end
         return EnzymeCore.BatchDuplicated(solution_val, solution_dval)
     end
@@ -66,7 +75,7 @@ function EnzymeRules.forward(
     end
 
     # downstream gradient
-    dz = ∂z∂θ * θ.dval
+    dz = ∂z∂θ_thunk * θ.dval
 
     solution_dval = (; solution_val..., z = dz)
 
@@ -77,4 +86,45 @@ function EnzymeRules.forward(
     @assert ReturnType <: EnzymeCore.Duplicated
 
     EnzymeCore.Duplicated(solution_val, solution_dval)
+end
+
+function ChainRulesCore.rrule(::typeof(ParametricMCPs.solve), problem, θ; kwargs...)
+    solution = ParametricMCPs.solve(problem, θ; kwargs...)
+    project_to_θ = ChainRulesCore.ProjectTo(θ)
+
+    function solve_pullback(∂solution)
+        no_grad_args = (; ∂self = ChainRulesCore.NoTangent(), ∂problem = ChainRulesCore.NoTangent())
+
+        ∂θ = ChainRulesCore.@thunk let
+            ∂z∂θ = _solve_jacobian_θ(problem, solution, θ)
+            ∂l∂z = ∂solution.z
+            project_to_θ(∂z∂θ' * ∂l∂z)
+        end
+
+        no_grad_args..., ∂θ
+    end
+
+    solution, solve_pullback
+end
+
+function ParametricMCPs.solve(
+    problem::ParametricMCPs.ParametricMCP,
+    θ::AbstractVector{<:ForwardDiff.Dual{T}};
+    kwargs...,
+) where {T}
+    # strip off the duals:
+    θ_v = ForwardDiff.value.(θ)
+    θ_p = ForwardDiff.partials.(θ)
+    # forward pass
+    solution = ParametricMCPs.solve(problem, θ_v; kwargs...)
+    # backward pass
+    ∂z∂θ = _solve_jacobian_θ(problem, solution, θ_v)
+    # downstream gradient
+    z_p = ∂z∂θ * θ_p
+    # glue forward and backward pass together into dual number types
+    z_d = ForwardDiff.Dual{T}.(solution.z, z_p)
+
+    (; z = z_d, solution.status, solution.info)
+end
+
 end
